refactor: contents dialog as a module
