+++ src/lib/src/contents-dialog/contents-dialog.component.scss
