+++ src/lib/src/contents-dialog/metadata/metadata.component.scss
